pyproject.toml
[tool.poetry]
name = "tw-picasso-diagnostic"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
psycopg = {extras = ["binary"], version = "^3.2.1"}
django-phonenumber-field = "^8.0.0"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.7"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
setuptools = "^70.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
